testing/web-platform/tests/resource-timing/buffer-full-when-populate-entries.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8" />
<link rel="author" title="Intel" href="http://www.intel.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#performanceresourcetiming"/>
<title>This test validates the functionality of onresourcetimingbufferfull in resource timing.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/webperftestharness.js"></script>
<script src="resources/webperftestharnessextension.js"></script>
<script src="resources/append.js"></script>
</head>
<body>
<script>
    const context = new PerformanceContext(performance);
    const t = async_test("Verify that resourcetimingbufferfull is properly invoked");
    performance.clearResourceTimings();
    let bufferSize = 2;
    context.setResourceTimingBufferSize(bufferSize);
    let bufferFullCount = 0;
    function buffer_full_callback(e) {
        assert_equals(e.bubbles, false, "Event bubbles attribute is false");
        bufferFullCount++;
    }
    context.registerResourceTimingBufferFullCallback(t.step_func(buffer_full_callback));
    // Scripts appended in JS to ensure setResourceTimingBufferSize is called before.
    appendScript('resources/empty.js');
    appendScript('resources/empty_script.js');
    appendScript('resources/resource_timing_test0.js');
    window.onload = t.step_func_done(()=>{
        assert_equals(context.getEntriesByType('resource').length, bufferSize, 'There should only be |bufferSize| resource entries.');
        assert_equals(bufferFullCount, 1, 'onresourcetimingbufferfull should have been invoked once.');
    });
</script>
</body>
</html>
